+++ src/main/java/com/api/base/domain/product/ProductDetailResponse.java
@@ -20,7 +20,7 @@ public class ProductDetailResponse {
     private String nameProduct;
     private ProductStatus status;
     private Double price;
-    private Double promotionalPrice;
+    private Integer promotionalPrice;
     private Integer quantity;
     private String description;
     private String use;
