+++ src/main/java/com/api/base/domain/product/ProductResponse.java
@@ -20,7 +20,7 @@ public class ProductResponse {
     private String nameProduct;
     private ProductStatus status;
     private Double price;
-    private Double promotionalPrice;
+    private Integer promotionalPrice;
     private Integer quantity;
     private String description;
     private String use;
